Moving scattered functions to more sensible locations.
