+++ ManuvrOS/DataStructures/Argument.h
@@ -2,6 +2,7 @@
 #define __MANUVR_ARGUMENT_H__
 
 #include <EnumeratedTypeCodes.h>
+#include <Types/TypeTranscriber.h>
 
 #include <DataStructures/Vector3.h>
 #include <DataStructures/Quaternion.h>
@@ -190,40 +191,5 @@ class Argument {
     // TODO: Might-should move this to someplace more accessable?
     static uintptr_t get_const_from_char_ptr(char*);
 };
-
-
-
-#if defined(MANUVR_CBOR)
-/* If we have CBOR support, we define a helper class to assist decomposition. */
-class CBORArgListener : public cbor::listener {
-  public:
-    CBORArgListener(Argument**);
-    ~CBORArgListener();
-
-    void on_integer(int value);
-    void on_bytes(unsigned char* data, int size);
-    void on_string(char* str);
-    void on_array(int size);
-    void on_map(int size);
-    void on_tag(unsigned int tag);
-    void on_special(unsigned int code);
-    void on_error(const char* error);
-
-    void on_extra_integer(unsigned long long value, int sign);
-    void on_extra_tag(unsigned long long tag);
-    void on_extra_special(unsigned long long tag);
-
-  private:
-    Argument** built = nullptr;
-    char*  _wait     = nullptr;
-    int _wait_map    = 0;
-    int _wait_array  = 0;
-
-    /* Please forgive the stupid name. */
-    void _caaa(Argument*);
-};
-#endif
-
-
 
 #endif  // __MANUVR_ARGUMENT_H__
