+++ ManuvrOS/Makefile
@@ -120,6 +120,7 @@ CPP_SRCS  += Drivers/INA219/INA219.cpp
 
 CPP_SRCS  += Platform/Cryptographic.cpp
 CPP_SRCS  += Platform/Identity.cpp
+CPP_SRCS  += Platform/Identity/IdentityUUID.cpp
 CPP_SRCS  += Platform/Storage.cpp
